Give compositor watchdog a module home
