+++ docs/development.md
@@ -51,12 +51,13 @@ then stops producing frames and the window looks frozen: unresponsive to clicks
 and keys even though the renderer and every process stay alive. It self-recovers
 after a few minutes, which is too long for a foreground app.
 
-`setupDarwinCompositorWatchdog` (`packages/desktop/src/window/window-manager.ts`)
-guards against this. It polls the renderer for frame production every couple of
-seconds and, after a sustained stall while the window is visible and unlocked,
-restarts the GPU process so Chromium rebuilds the display link. The probe is
-skipped while the screen is locked or the window is hidden or minimized, since a
-window legitimately stops producing frames then.
+`setupDarwinCompositorWatchdog`
+(`packages/desktop/src/window/compositor-watchdog/index.ts`) guards against
+this. It polls the renderer for frame production every couple of seconds and,
+after a sustained stall while the window is visible and unlocked, restarts the
+GPU process so Chromium rebuilds the display link. The probe is skipped while
+the screen is locked or the window is hidden or minimized, since a window
+legitimately stops producing frames then.
 
 ### Daemon logs
